Added PartitionCacheTest. Changed MessageCache into an interface for testability.
